Refactoring code after review
