+++ TeletekstBot.Infrastructure/Services/MastodonService.cs
@@ -3,7 +3,6 @@ using Mastonet.Entities;
 using Microsoft.Extensions.Logging;
 using TeletekstBot.Application.Interfaces;
 using TeletekstBot.Domain.Entities;
-using TeletekstBot.Infrastructure.Extensions;
 
 namespace TeletekstBot.Infrastructure.Services;
 
@@ -20,7 +19,7 @@ public class MastodonService : IMastodonService
     {
         _logger.LogInformation("[Mastodon] Posting page {PageNr} with title '{Title}'", page.PageNumber, page.Title);
 
-        var attachment = await _mastodonClient.UploadMedia(screenshotStream, "screenshot.jpg", page.Body.AddSpacesWhenApplicable(), new AttachmentFocusData
+        var attachment = await _mastodonClient.UploadMedia(screenshotStream, "screenshot.jpg", page.Body, new AttachmentFocusData
         {
             X = 0.0,
             Y = 1.0
